move auth stuff to express
